requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt
django-cors-headers
gunicorn>=21.2
psycopg2-binary>=2.9
redis>=5.0
django-environ>=0.11
celery>=5.3
django-cors-headers>=4.3
stripe>=7.0
Pillow>=10.0
python-decouple
dj-database-url